refactor: WaitForAsync as first class citizen
