+++ src/bunit/Extensions/WaitForHelpers/RenderedFragmentWaitForHelperExtensions.cs
@@ -10,30 +10,6 @@ namespace Bunit;
 /// </summary>
 public static class RenderedFragmentWaitForHelperExtensions
 {
-	/// <summary>
-	/// Wait until the provided <paramref name="statePredicate"/> action returns true,
-	/// or the <paramref name="timeout"/> is reached (default is one second).
-	///
-	/// The <paramref name="statePredicate"/> is evaluated initially, and then each time
-	/// the <paramref name="renderedFragment"/> renders.
-	/// </summary>
-	/// <param name="renderedFragment">The render fragment or component to attempt to verify state against.</param>
-	/// <param name="statePredicate">The predicate to invoke after each render, which must returns <c>true</c> when the desired state has been reached.</param>
-	/// <param name="timeout">The maximum time to wait for the desired state.</param>
-	/// <exception cref="WaitForFailedException">Thrown if the <paramref name="statePredicate"/> throw an exception during invocation, or if the timeout has been reached. See the inner exception for details.</exception>
-	public static void WaitForState(this IRenderedFragment renderedFragment, Func<bool> statePredicate, TimeSpan? timeout = null)
-	{
-		using var waiter = new WaitForStateHelper(renderedFragment, statePredicate, timeout);
-
-		try
-		{
-			waiter.WaitTask.GetAwaiter().GetResult();
-		}
-		catch (AggregateException e) when (e.InnerExceptions.Count == 1)
-		{
-			ExceptionDispatchInfo.Capture(e.InnerExceptions[0]).Throw();
-		}
-	}
 
 	/// <summary>
 	/// Wait until the provided <paramref name="statePredicate"/> action returns true,
@@ -46,7 +22,7 @@ public static class RenderedFragmentWaitForHelperExtensions
 	/// <param name="statePredicate">The predicate to invoke after each render, which must returns <c>true</c> when the desired state has been reached.</param>
 	/// <param name="timeout">The maximum time to wait for the desired state.</param>
 	/// <exception cref="WaitForFailedException">Thrown if the <paramref name="statePredicate"/> throw an exception during invocation, or if the timeout has been reached. See the inner exception for details.</exception>
-	internal static async Task WaitForStateAsync(this IRenderedFragment renderedFragment, Func<bool> statePredicate, TimeSpan? timeout = null)
+	public static async Task WaitForStateAsync(this IRenderedFragment renderedFragment, Func<bool> statePredicate, TimeSpan? timeout = null)
 	{
 		using var waiter = new WaitForStateHelper(renderedFragment, statePredicate, timeout);
 
@@ -64,32 +40,7 @@ public static class RenderedFragmentWaitForHelperExtensions
 	/// <param name="timeout">The maximum time to attempt the verification.</param>
 	/// <exception cref="WaitForFailedException">Thrown if the timeout has been reached. See the inner exception to see the captured assertion exception.</exception>
 	[AssertionMethod]
-	public static void WaitForAssertion(this IRenderedFragment renderedFragment, Action assertion, TimeSpan? timeout = null)
-	{
-		using var waiter = new WaitForAssertionHelper(renderedFragment, assertion, timeout);
-
-		try
-		{
-			waiter.WaitTask.GetAwaiter().GetResult();
-		}
-		catch (AggregateException e) when (e.InnerExceptions.Count == 1)
-		{
-			ExceptionDispatchInfo.Capture(e.InnerExceptions[0]).Throw();
-		}
-	}
-
-	/// <summary>
-	/// Wait until the provided <paramref name="assertion"/> passes (i.e. does not throw an
-	/// exception), or the <paramref name="timeout"/> is reached (default is one second).
-	///
-	/// The <paramref name="assertion"/> is attempted initially, and then each time the <paramref name="renderedFragment"/> renders.
-	/// </summary>
-	/// <param name="renderedFragment">The rendered fragment to wait for renders from and assert against.</param>
-	/// <param name="assertion">The verification or assertion to perform.</param>
-	/// <param name="timeout">The maximum time to attempt the verification.</param>
-	/// <exception cref="WaitForFailedException">Thrown if the timeout has been reached. See the inner exception to see the captured assertion exception.</exception>
-	[AssertionMethod]
-	internal static async Task WaitForAssertionAsync(this IRenderedFragment renderedFragment, Action assertion, TimeSpan? timeout = null)
+	public static async Task WaitForAssertionAsync(this IRenderedFragment renderedFragment, Action assertion, TimeSpan? timeout = null)
 	{
 		using var waiter = new WaitForAssertionHelper(renderedFragment, assertion, timeout);
 
@@ -104,78 +55,7 @@ public static class RenderedFragmentWaitForHelperExtensions
 	/// <param name="cssSelector">The CSS selector to use to search for the element.</param>
 	/// <exception cref="WaitForFailedException">Thrown if no elements is found matching the <paramref name="cssSelector"/> within the default timeout. See the inner exception for details.</exception>
 	/// <returns>The <see cref="IElement"/>.</returns>
-	public static IElement WaitForElement(this IRenderedFragment renderedFragment, string cssSelector)
-		=> WaitForElementCore(renderedFragment, cssSelector, timeout: null);
-
-	/// <summary>
-	/// Wait until an element matching the <paramref name="cssSelector"/> exists in the <paramref name="renderedFragment"/>,
-	/// or the <paramref name="timeout"/> is reached.
-	/// </summary>
-	/// <param name="renderedFragment">The render fragment or component find the matching element in.</param>
-	/// <param name="cssSelector">The CSS selector to use to search for the element.</param>
-	/// <param name="timeout">The maximum time to wait for the element to appear.</param>
-	/// <exception cref="WaitForFailedException">Thrown if no elements is found matching the <paramref name="cssSelector"/> within the default timeout. See the inner exception for details.</exception>
-	/// <returns>The <see cref="IElement"/>.</returns>
-	public static IElement WaitForElement(this IRenderedFragment renderedFragment, string cssSelector, TimeSpan timeout)
-		=> WaitForElementCore(renderedFragment, cssSelector, timeout: timeout);
-
-	/// <summary>
-	/// Wait until at least one element matching the <paramref name="cssSelector"/> exists in the <paramref name="renderedFragment"/>,
-	/// or the timeout is reached (default is one second).
-	/// </summary>
-	/// <param name="renderedFragment">The render fragment or component find the matching element in.</param>
-	/// <param name="cssSelector">The CSS selector to use to search for elements.</param>
-	/// <exception cref="WaitForFailedException">Thrown if no elements is found matching the <paramref name="cssSelector"/> within the default timeout.</exception>
-	/// <returns>The <see cref="IRefreshableElementCollection{IElement}"/>.</returns>
-	public static IRefreshableElementCollection<IElement> WaitForElements(this IRenderedFragment renderedFragment, string cssSelector)
-		=> WaitForElementsCore(renderedFragment, cssSelector, matchElementCount: null, timeout: null);
-
-	/// <summary>
-	/// Wait until exactly <paramref name="matchElementCount"/> element(s) matching the <paramref name="cssSelector"/> exists in the <paramref name="renderedFragment"/>,
-	/// or the timeout is reached (default is one second).
-	/// </summary>
-	/// <param name="renderedFragment">The render fragment or component find the matching element in.</param>
-	/// <param name="cssSelector">The CSS selector to use to search for elements.</param>
-	/// <param name="matchElementCount">The exact number of elements to that the <paramref name="cssSelector"/> should match.</param>
-	/// <exception cref="WaitForFailedException">Thrown if no elements is found matching the <paramref name="cssSelector"/> within the default timeout.</exception>
-	/// <returns>The <see cref="IRefreshableElementCollection{IElement}"/>.</returns>
-	public static IRefreshableElementCollection<IElement> WaitForElements(this IRenderedFragment renderedFragment, string cssSelector, int matchElementCount)
-		=> WaitForElementsCore(renderedFragment, cssSelector, matchElementCount: matchElementCount, timeout: null);
-
-	/// <summary>
-	/// Wait until at least one element matching the <paramref name="cssSelector"/> exists in the <paramref name="renderedFragment"/>,
-	/// or the <paramref name="timeout"/> is reached.
-	/// </summary>
-	/// <param name="renderedFragment">The render fragment or component find the matching element in.</param>
-	/// <param name="cssSelector">The CSS selector to use to search for elements.</param>
-	/// <param name="timeout">The maximum time to wait for elements to appear.</param>
-	/// <exception cref="WaitForFailedException">Thrown if no elements is found matching the <paramref name="cssSelector"/> within the default timeout.</exception>
-	/// <returns>The <see cref="IRefreshableElementCollection{IElement}"/>.</returns>
-	public static IRefreshableElementCollection<IElement> WaitForElements(this IRenderedFragment renderedFragment, string cssSelector, TimeSpan timeout)
-		=> WaitForElementsCore(renderedFragment, cssSelector, matchElementCount: null, timeout: timeout);
-
-	/// <summary>
-	/// Wait until exactly <paramref name="matchElementCount"/> element(s) matching the <paramref name="cssSelector"/> exists in the <paramref name="renderedFragment"/>,
-	/// or the <paramref name="timeout"/> is reached.
-	/// </summary>
-	/// <param name="renderedFragment">The render fragment or component find the matching element in.</param>
-	/// <param name="cssSelector">The CSS selector to use to search for elements.</param>
-	/// <param name="matchElementCount">The exact number of elements to that the <paramref name="cssSelector"/> should match.</param>
-	/// <param name="timeout">The maximum time to wait for elements to appear.</param>
-	/// <exception cref="WaitForFailedException">Thrown if no elements is found matching the <paramref name="cssSelector"/> within the default timeout.</exception>
-	/// <returns>The <see cref="IRefreshableElementCollection{IElement}"/>.</returns>
-	public static IRefreshableElementCollection<IElement> WaitForElements(this IRenderedFragment renderedFragment, string cssSelector, int matchElementCount, TimeSpan timeout)
-		=> WaitForElementsCore(renderedFragment, cssSelector, matchElementCount: matchElementCount, timeout: timeout);
-
-	/// <summary>
-	/// Wait until an element matching the <paramref name="cssSelector"/> exists in the <paramref name="renderedFragment"/>,
-	/// or the timeout is reached (default is one second).
-	/// </summary>
-	/// <param name="renderedFragment">The render fragment or component find the matching element in.</param>
-	/// <param name="cssSelector">The CSS selector to use to search for the element.</param>
-	/// <exception cref="WaitForFailedException">Thrown if no elements is found matching the <paramref name="cssSelector"/> within the default timeout. See the inner exception for details.</exception>
-	/// <returns>The <see cref="IElement"/>.</returns>
-	internal static Task<IElement> WaitForElementAsync(this IRenderedFragment renderedFragment, string cssSelector)
+	public static Task<IElement> WaitForElementAsync(this IRenderedFragment renderedFragment, string cssSelector)
 		=> WaitForElementCoreAsync(renderedFragment, cssSelector, timeout: null);
 
 	/// <summary>
@@ -235,52 +115,14 @@ public static class RenderedFragmentWaitForHelperExtensions
 	/// <param name="cssSelector">The CSS selector to use to search for elements.</param>
 	/// <exception cref="WaitForFailedException">Thrown if no elements is found matching the <paramref name="cssSelector"/> within the default timeout.</exception>
 	/// <returns>The <see cref="IRefreshableElementCollection{IElement}"/>.</returns>
-	internal static Task<IRefreshableElementCollection<IElement>> WaitForElementsAsync(this IRenderedFragment renderedFragment, string cssSelector)
+	public static Task<IRefreshableElementCollection<IElement>> WaitForElementsAsync(this IRenderedFragment renderedFragment, string cssSelector)
 		=> WaitForElementsCoreAsync(renderedFragment, cssSelector, matchElementCount: null, timeout: null);
-
-	private static IElement WaitForElementCore(this IRenderedFragment renderedFragment, string cssSelector, TimeSpan? timeout)
-	{
-		using var waiter = new WaitForElementHelper(renderedFragment, cssSelector, timeout);
-
-		try
-		{
-			return waiter.WaitTask.GetAwaiter().GetResult();
-		}
-		catch (AggregateException e) when (e.InnerExceptions.Count == 1)
-		{
-			ExceptionDispatchInfo.Capture(e.InnerExceptions[0]).Throw();
-
-			// Unreachable code
-			throw;
-		}
-	}
 
 	private static async Task<IElement> WaitForElementCoreAsync(this IRenderedFragment renderedFragment, string cssSelector, TimeSpan? timeout)
 	{
 		using var waiter = new WaitForElementHelper(renderedFragment, cssSelector, timeout);
 
 		return await waiter.WaitTask;
-	}
-
-	private static IRefreshableElementCollection<IElement> WaitForElementsCore(
-		this IRenderedFragment renderedFragment,
-		string cssSelector,
-		int? matchElementCount,
-		TimeSpan? timeout)
-	{
-		using var waiter = new WaitForElementsHelper(renderedFragment, cssSelector, matchElementCount, timeout);
-
-		try
-		{
-			return waiter.WaitTask.GetAwaiter().GetResult();
-		}
-		catch (AggregateException e) when (e.InnerExceptions.Count == 1)
-		{
-			ExceptionDispatchInfo.Capture(e.InnerExceptions[0]).Throw();
-
-			// Unreachable code
-			throw;
-		}
 	}
 
 	private static async Task<IRefreshableElementCollection<IElement>> WaitForElementsCoreAsync(
