+++ tests/bunit.tests/TestDoubles/Authorization/AuthorizationTest.cs
@@ -52,7 +52,6 @@ public class AuthorizationTest : TestContext
 	}
 
 	[Fact(DisplayName = "AuthorizeView switch from unauthorized to authorized.")]
-	[Trait("Category", "async")]
 	public async Task Test004()
 	{
 		// arrange
@@ -69,26 +68,6 @@ public class AuthorizationTest : TestContext
 
 		// assert
 		await cut.WaitForAssertionAsync(() => cut.MarkupMatches("Authorized!"));
-	}
-
-	[Fact(DisplayName = "AuthorizeView switch from unauthorized to authorized.")]
-	[Trait("Category", "sync")]
-	public void Test004_Sync()
-	{
-		// arrange
-		var authContext = this.AddTestAuthorization();
-
-		// start off unauthenticated.
-		var cut = RenderComponent<SimpleAuthView>();
-		cut.MarkupMatches("Not authorized?");
-
-		// act
-		authContext.SetAuthorized("TestUser004", AuthorizationState.Authorized);
-
-		cut.Render();
-
-		// assert
-		cut.WaitForAssertion(() => cut.MarkupMatches("Authorized!"));
 	}
 
 	[Fact(DisplayName = "AuthorizeView switch from authorized to unauthorized.")]
